+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.2] - Unreleased
+### Fixed
+- Forward-port changes from release-1.0 (gracefully handle malformed
+  purge properties messages).
+
 ## [1.1.1.] - 2023-11-15
 ### Fixed
 - [astarte_data_updater_plant] Don't crash when retrieving the interface version
@@ -59,7 +64,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   which are not fully populated. Fix [531](astarte-platform#531).
 
 ## [1.0.6] - Unreleased
+### Fixed
 - [astarte_appengine_api] Allow to send binaryblobarrays over server owned interfaces.
+- [astarte_data_updater_plant] Do not crash when receiving a malformed purge properties message.
 
 ## [1.0.5] - 2023-09-26
 ### Fixed